lrem: refactor build
Create separate libraries for the db and mem modules.
Rename: librados_stub -> librados_sqlite, notrados -> rados-sqlite.
Also add separate radosgw-sqlite, radosgw-admin-sqlite.
