+++ ceph.spec.in
@@ -1582,8 +1582,9 @@ exit 0
 %{_bindir}/cephfs-table-tool
 %{_bindir}/crushdiff
 %{_bindir}/rados
-%{_bindir}/notrados
+%{_bindir}/rados-sqlite
 %{_bindir}/radosgw-admin
+%{_bindir}/radosgw-admin-sqlite
 %{_bindir}/rbd
 %{_bindir}/rbd-replay
 %{_bindir}/rbd-replay-many
@@ -2038,6 +2039,7 @@ fi
 %files radosgw
 %{_bindir}/ceph-diff-sorted
 %{_bindir}/radosgw
+%{_bindir}/radosgw-sqlite
 %{_bindir}/radosgw-token
 %{_bindir}/radosgw-es
 %{_bindir}/radosgw-object-expirer
@@ -2202,7 +2204,8 @@ fi
 
 %files -n librados2
 %{_libdir}/librados.so.*
-%{_libdir}/librados_stub.so.*
+%{_libdir}/librados_sqlite.so.*
+%{_libdir}/librados_mem.so.*
 %dir %{_libdir}/ceph
 %{_libdir}/ceph/libceph-common.so.*
 %if %{with lttng}
@@ -2219,7 +2222,8 @@ fi
 %{_includedir}/rados/librados.h
 %{_includedir}/rados/rados_types.h
 %{_libdir}/librados.so
-%{_libdir}/librados_stub.so
+%{_libdir}/librados_sqlite.so
+%{_libdir}/librados_mem.so
 %if %{with lttng}
 %{_libdir}/librados_tp.so
 %endif
