+++ src/tools/CMakeLists.txt
@@ -16,14 +16,14 @@ else()
 endif()
 install(TARGETS rados DESTINATION bin)
 
-add_executable(notrados ${rados_srcs})
-target_link_libraries(notrados librados_stub global ${BLKID_LIBRARIES} ${CMAKE_DL_LIBS})
+add_executable(rados-sqlite ${rados_srcs})
+target_link_libraries(rados-sqlite librados_sqlite global ${BLKID_LIBRARIES} ${CMAKE_DL_LIBS})
 if(WITH_LIBRADOSSTRIPER)
-  target_link_libraries(notrados radosstriper)
+  target_link_libraries(rados-sqlite radosstriper)
 else()
-  target_link_libraries(notrados cls_lock_client)
+  target_link_libraries(rados-sqlite cls_lock_client)
 endif()
-install(TARGETS notrados DESTINATION bin)
+install(TARGETS rados-sqlite DESTINATION bin)
 
 if(NOT WIN32)
   set(neorados_srcs
